+++ src/periodic_table.h
@@ -12,11 +12,11 @@
 
 class PeriodicTable {
     std::vector<Element> elements_;
-    std::map<std::string, int> symbol_Z_;
+    std::map<std::string, size_t> symbol_Z_;
 public:
     static const PeriodicTable &pte();
 
-    const Element *getElement(int Z) const { return &elements_[Z]; }
+    const Element *getElement(size_t Z) const { return &elements_[Z]; }
 
     const Element *getElement(const std::string &symbol) const;
